+++ src/types/CartItem.type.ts
@@ -1,0 +1,7 @@
+import { TProduct } from './Product.type';
+
+export type TCartItem = {
+  id: number;
+  quantity: number;
+  product: TProduct;
+};
